refactor: decouple Jira integration into dojo/jira package
Introduce a new dojo/jira package (models, forms, api, helper, queries,
urls, views) and a top-level dojo/jira_facade.py. Route all Jira
references from the rest of the codebase (models.py properties, forms,
api_v2 serializers and views, view files, tasks, display_tags,
risk_acceptance/helper, finding/helper, importers, utils SLA
notifications, engagement/services, management commands) through the
facade so no module outside the Jira package imports Jira internals
directly.

Also resolves a circular import in dojo/jira/models.py, updates test
mock paths for the new module layout, and fixes ruff lint across the
moved files.
