+++ dojo/finding_group/views.py
@@ -12,7 +12,6 @@ from django.urls.base import reverse
 from django.views import View
 from django.views.decorators.http import require_POST
 
-import dojo.jira_link.helper as jira_helper
 from dojo.authorization.authorization import user_has_permission_or_403
 from dojo.authorization.authorization_decorators import user_is_authorized
 from dojo.authorization.roles_permissions import Permissions
@@ -23,6 +22,7 @@ from dojo.filters import (
 )
 from dojo.finding.queries import prefetch_for_findings
 from dojo.forms import DeleteFindingGroupForm, EditFindingGroupForm, FindingBulkUpdateForm
+from dojo.jira import services as jira_services
 from dojo.models import Engagement, Finding, Finding_Group, GITHUB_PKey, Global_Role, Product
 from dojo.product.queries import get_authorized_products
 from dojo.utils import Product_Tab, add_breadcrumb, get_page_items, get_setting, get_system_setting, get_words_for_field
@@ -48,7 +48,7 @@ def view_finding_group(request, fgid):
         product = get_object_or_404(Product, id=pid)
         user_has_permission_or_403(request.user, product, Permissions.Product_View)
         product_tab = Product_Tab(product, title="Findings", tab="findings")
-        jira_project = jira_helper.get_jira_project(product)
+        jira_project = jira_services.get_project(product)
         github_config = GITHUB_PKey.objects.filter(product=pid).first()
         findings_filter = finding_filter_class(request.GET, findings, user=request.user, pid=pid)
     elif finding_group.test.engagement.id:
@@ -56,7 +56,7 @@ def view_finding_group(request, fgid):
         engagement = get_object_or_404(Engagement, id=eid)
         user_has_permission_or_403(request.user, engagement, Permissions.Engagement_View)
         product_tab = Product_Tab(engagement.product, title=engagement.name, tab="engagements")
-        jira_project = jira_helper.get_jira_project(engagement)
+        jira_project = jira_services.get_project(engagement)
         github_config = GITHUB_PKey.objects.filter(product__engagement=eid).first()
         findings_filter = finding_filter_class(request.GET, findings, user=request.user, eid=eid)
 
@@ -82,7 +82,7 @@ def view_finding_group(request, fgid):
 
             if jira_issue:
                 # See if the submitted issue was a issue key or the full URL
-                jira_project = jira_helper.get_jira_project(finding_group)
+                jira_project = jira_services.get_project(finding_group)
                 if not jira_project or not jira_project.jira_instance:
                     messages.add_message(
                         request,
@@ -94,13 +94,13 @@ def view_finding_group(request, fgid):
                 jira_instance = jira_project.jira_instance
                 jira_issue = jira_issue.removeprefix(jira_instance.url + "/browse/")
 
-                if finding_group.has_jira_issue and jira_issue != jira_helper.get_jira_key(finding_group):
-                    jira_helper.unlink_jira(request, finding_group)
-                    jira_helper.finding_group_link_jira(request, finding_group, jira_issue)
+                if finding_group.has_jira_issue and jira_issue != jira_services.get_key(finding_group):
+                    jira_services.unlink(request, finding_group)
+                    jira_services.link_finding_group(request, finding_group, jira_issue)
                 elif not finding_group.has_jira_issue:
-                    jira_helper.finding_group_link_jira(request, finding_group, jira_issue)
+                    jira_services.link_finding_group(request, finding_group, jira_issue)
             elif push_to_jira:
-                jira_helper.push_to_jira(finding_group, sync=True)
+                jira_services.push(finding_group, sync=True)
 
             finding_group.save()
             return HttpResponseRedirect(reverse("view_test", args=(finding_group.test.id,)))
@@ -162,7 +162,7 @@ def unlink_jira(request, fgid):
     logger.info("trying to unlink a linked jira issue from %d:%s", group.id, group.name)
     if group.has_jira_issue:
         try:
-            jira_helper.unlink_jira(request, group)
+            jira_services.unlink(request, group)
 
             messages.add_message(
                 request,
@@ -200,7 +200,7 @@ def push_to_jira(request, fgid):
 
         # it may look like success here, but the push_to_jira are swallowing exceptions
         # but cant't change too much now without having a test suite, so leave as is for now with the addition warning message to check alerts for background errors.
-        if jira_helper.push_to_jira(group, sync=True):
+        if jira_services.push(group, sync=True):
             messages.add_message(
                 request,
                 messages.SUCCESS,
