+++ dojo/importers/default_reimporter.py
@@ -6,7 +6,6 @@ from django.core.files.uploadedfile import TemporaryUploadedFile
 from django.db.models.query_utils import Q
 
 import dojo.finding.helper as finding_helper
-import dojo.jira_link.helper as jira_helper
 from dojo.celery_dispatch import dojo_dispatch_task
 from dojo.finding.deduplication import (
     find_candidates_for_deduplication_hash,
@@ -17,7 +16,7 @@ from dojo.finding.deduplication import (
 from dojo.importers.base_importer import BaseImporter, Parser
 from dojo.importers.endpoint_manager import EndpointManager
 from dojo.importers.options import ImporterOptions
-from dojo.jira_link.helper import is_keep_in_sync_with_jira
+from dojo.jira import services as jira_services
 from dojo.location.status import FindingLocationStatus
 from dojo.models import (
     Development_Environment,
@@ -564,8 +563,8 @@ class DefaultReImporter(BaseImporter, DefaultReImporterOptions):
         if self.findings_groups_enabled and (self.push_to_jira or getattr(self.jira_instance, "finding_jira_sync", False)):
             for finding_group in {finding.finding_group for finding in findings if finding.finding_group is not None}:
                 # Check the push_to_jira flag again to potentially shorty circuit without checking for existing findings
-                if self.push_to_jira or is_keep_in_sync_with_jira(finding_group, prefetched_jira_instance=self.jira_instance):
-                    jira_helper.push_to_jira(finding_group)
+                if self.push_to_jira or jira_services.is_keep_in_sync(finding_group, prefetched_jira_instance=self.jira_instance):
+                    jira_services.push(finding_group)
         # Calculate grade once after all findings have been closed
         if mitigated_findings:
             perform_product_grading(self.test.engagement.product)
@@ -1065,8 +1064,8 @@ class DefaultReImporter(BaseImporter, DefaultReImporterOptions):
             if self.push_to_jira or getattr(self.jira_instance, "finding_jira_sync", False):
                 object_to_push = findings[0].finding_group if findings[0].finding_group is not None else findings[0]
                 # Check the push_to_jira flag again to potentially shorty circuit without checking for existing findings
-                if self.push_to_jira or is_keep_in_sync_with_jira(object_to_push, prefetched_jira_instance=self.jira_instance):
-                    jira_helper.push_to_jira(object_to_push)
+                if self.push_to_jira or jira_services.is_keep_in_sync(object_to_push, prefetched_jira_instance=self.jira_instance):
+                    jira_services.push(object_to_push)
         # We dont check if the finding jira sync is applicable quite yet until we can get in the loop
         # but this is a way to at least make it that far
         if self.findings_groups_enabled and (self.push_to_jira or getattr(self.jira_instance, "finding_jira_sync", False)):
@@ -1076,8 +1075,8 @@ class DefaultReImporter(BaseImporter, DefaultReImporterOptions):
                     if finding.finding_group is not None and not finding.is_mitigated
             }:
                 # Check the push_to_jira flag again to potentially shorty circuit without checking for existing findings
-                if self.push_to_jira or is_keep_in_sync_with_jira(finding_group, prefetched_jira_instance=self.jira_instance):
-                    jira_helper.push_to_jira(finding_group)
+                if self.push_to_jira or jira_services.is_keep_in_sync(finding_group, prefetched_jira_instance=self.jira_instance):
+                    jira_services.push(finding_group)
 
     def calculate_unsaved_finding_hash_code(
         self,
