+++ dojo/management/commands/jira_refactor_data_migration.py
@@ -2,7 +2,7 @@ import logging
 
 from django.core.management.base import BaseCommand
 
-import dojo.jira_link.helper as jira_helper
+import dojo.jira.helper as jira_helper
 from dojo.models import JIRA_Instance, JIRA_Issue
 
 logger = logging.getLogger(__name__)
