+++ dojo/tasks.py
@@ -17,7 +17,11 @@ from dojo.celery import app
 from dojo.celery_dispatch import dojo_dispatch_task
 from dojo.finding.helper import fix_loop_duplicates
 from dojo.location.models import Location
-from dojo.management.commands.jira_status_reconciliation import jira_status_reconciliation
+
+try:
+    from dojo.management.commands.jira_status_reconciliation import jira_status_reconciliation
+except ImportError:
+    jira_status_reconciliation = None
 from dojo.models import Alerts, Announcement, Endpoint, Engagement, Finding, Product, System_Settings, User
 from dojo.notifications.helper import create_notification
 from dojo.utils import calculate_grade, sla_compute_and_notify
@@ -203,6 +207,9 @@ def async_sla_compute_and_notify_task(*args, **kwargs):
 
 @app.task
 def jira_status_reconciliation_task(*args, **kwargs):
+    if jira_status_reconciliation is None:
+        logger.warning("Jira status reconciliation is not available")
+        return None
     # Wrap with pghistory context for audit trail
     with pghistory.context(
         source="jira_reconciliation",
