+++ dojo/templatetags/display_tags.py
@@ -24,9 +24,9 @@ from django.utils.html import conditional_escape, escape
 from django.utils.safestring import mark_safe
 from django.utils.translation import gettext as _
 
-import dojo.jira_link.helper as jira_helper
 import dojo.utils
 from dojo import __docs__, __version__
+from dojo.jira import services as jira_services
 from dojo.models import Benchmark_Product, Check_List, Dojo_User, FileAccessToken, Finding, Product, System_Settings
 from dojo.utils import calculate_grade, get_file_images, get_full_url, get_system_setting, prepare_for_view
 
@@ -909,52 +909,52 @@ def jiraencode_component(value):
 
 @register.filter
 def jira_project(obj, *, use_inheritance=True):
-    return jira_helper.get_jira_project(obj, use_inheritance=use_inheritance)
+    return jira_services.get_project(obj, use_inheritance=use_inheritance)
 
 
 @register.filter
 def jira_issue_url(obj):
-    return jira_helper.get_jira_url(obj)
+    return jira_services.get_url(obj)
 
 
 @register.filter
 def jira_project_url(obj):
-    return jira_helper.get_jira_project_url(obj)
+    return jira_services.get_project_url(obj)
 
 
 @register.filter
 def jira_key(obj):
-    return jira_helper.get_jira_key(obj)
+    return jira_services.get_key(obj)
 
 
 @register.filter
 def jira_creation(obj):
-    return jira_helper.get_jira_creation(obj)
+    return jira_services.get_creation(obj)
 
 
 @register.filter
 def jira_change(obj):
-    return jira_helper.get_jira_change(obj)
+    return jira_services.get_change(obj)
 
 
 @register.filter
 def jira_qualified_findings(finding_group):
-    return jira_helper.get_qualified_findings(finding_group)
+    return jira_services.get_qualified_findings(finding_group)
 
 
 @register.filter
 def jira_non_qualified_findings(finding_group):
-    return jira_helper.get_non_qualified_findings(finding_group)
+    return jira_services.get_non_qualified_findings(finding_group)
 
 
 @register.filter
 def jira_sla_deadline(obj):
-    return jira_helper.get_sla_deadline(obj)
+    return jira_services.get_sla_deadline(obj)
 
 
 @register.filter
 def jira_severity(findings):
-    return jira_helper.get_severity(findings)
+    return jira_services.get_severity(findings)
 
 
 @register.filter
@@ -1012,7 +1012,7 @@ def jira_project_tag(product_or_engagement, *, autoescape=True):
         def esc(x):
             return x
 
-    jira_project = jira_helper.get_jira_project(product_or_engagement)
+    jira_project = jira_services.get_project(product_or_engagement)
 
     if not jira_project:
         return ""
@@ -1028,7 +1028,7 @@ def jira_project_tag(product_or_engagement, *, autoescape=True):
         <b>Push Notes:</b> %s">
     </i>
     """
-    jira_project_no_inheritance = jira_helper.get_jira_project(product_or_engagement, use_inheritance=False)
+    jira_project_no_inheritance = jira_services.get_project(product_or_engagement, use_inheritance=False)
     inherited = bool(not jira_project_no_inheritance)
 
     icon = "fa-bug"
