+++ unittests/test_jira_webhook.py
@@ -5,7 +5,7 @@ import logging
 
 from django.urls import reverse
 
-import dojo.jira_link.helper as jira_helper
+import dojo.jira.helper as jira_helper
 from dojo.models import JIRA_Issue
 from unittests.dojo_test_case import DojoTestCase, versioned_fixtures
